requirements.txt
disnake[voice]
requests
youtube_dl
Pillow
numpy
scipy
PyNaCl~=1.4.0